# internal/api/handler.go
package api

import (
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	httpSwagger "github.com/swaggo/http-swagger/v2"
	_ "github.com/user/whatsmeow-basileia/docs" // Docs generated by swaggo
	"github.com/user/whatsmeow-basileia/internal/infrastructure/rabbitmq"
	"github.com/user/whatsmeow-basileia/internal/infrastructure/whatsapp"
	"github.com/user/whatsmeow-basileia/pkg/logger"
	"go.uber.org/zap"
)

type SendMessageRequest struct {
	DeviceID   string   `json:"device_id,omitempty"`
	Number     string   `json:"number,omitempty"`
	Phone      string   `json:"phone,omitempty" swaggerignore:"true"`     // Uazapi alias for Number
	Message    string   `json:"message,omitempty"`
	Text       string   `json:"text,omitempty" swaggerignore:"true"`       // Uazapi alias for Message
	Footer     string   `json:"footer,omitempty"`
	FooterText string   `json:"footerText,omitempty" swaggerignore:"true"` // Uazapi alias for Footer
	Title      string   `json:"title,omitempty"`
	Type       string   `json:"type,omitempty"`       // button, list
	Buttons    []string `json:"buttons,omitempty"`
	Choices    []string `json:"choices,omitempty" swaggerignore:"true"`    // Uazapi alias for Buttons
	ListButton string   `json:"listButton,omitempty" swaggerignore:"true"` // for list button text
}

type SendMediaRequest struct {
	DeviceID   string `json:"device_id"`
	Number     string `json:"number,omitempty"`
	Phone      string `json:"phone,omitempty" swaggerignore:"true"` // Uazapi alias for Number
	MediaURL   string `json:"media_url,omitempty"` // Public URL to download the media from
	Base64     string `json:"base64,omitempty"`    // Base64 encoded media string (e.g. data:image/png;base64,...)
	Type       string `json:"type"`                // image, video, document, audio
	Caption    string `json:"caption,omitempty"`
	Text       string `json:"text,omitempty" swaggerignore:"true"` // Uazapi alias for Caption
	MimeType   string `json:"mimetype,omitempty"`
	FileName   string `json:"fileName,omitempty"`
}

type SendCopyButtonRequest struct {
	DeviceID string `json:"device_id"`
	Number   string `json:"number,omitempty"`
	Phone    string `json:"phone,omitempty" swaggerignore:"true"` // Alias
	Message  string `json:"message,omitempty"`
	Text     string `json:"text,omitempty" swaggerignore:"true"` // Alias
	Title    string `json:"title,omitempty"`
	Footer   string `json:"footer,omitempty"`
	CopyCode string `json:"copy_code"` // The actual content to be copied (e.g. PIX payload)
	CopyText string `json:"copy_text"` // The button display text (e.g. "Copy PIX")
}

type Handler struct {
	manager   *whatsapp.MultiClientManager
	rmqClient *rabbitmq.Client
}

func NewHandler(manager *whatsapp.MultiClientManager, rmqClient *rabbitmq.Client) *Handler {
	return &Handler{
		manager:   manager,
		rmqClient: rmqClient,
	}
}

func (h *Handler) Router() http.Handler {
	r := chi.NewRouter()
	r.Use(middleware.Logger)

	r.Post("/device", h.CreateDevice)
	r.Get("/device", h.ListDevices)
	r.Put("/device/{id}/rename", h.RenameDevice)
	r.Delete("/device/{id}", h.DeleteDevice)
	r.Post("/device/{id}/reconnect", h.ReconnectDevice)
	r.Get("/device/{id}/qr", h.GetQR)
	r.Get("/device/{id}/status", h.GetStatus)
	r.Post("/message/send", h.SendText)
	r.Post("/message/interactive/button", h.SendInteractiveButton)
	r.Post("/message/interactive/copy", h.SendInteractiveCopyButton)
	r.Post("/message/media", h.SendMedia)
	r.Get("/logs", h.GetLogs)
	r.Get("/stats", h.GetStats)

	r.Get("/docs/*", httpSwagger.Handler(
		httpSwagger.URL("/docs/doc.json"), //The url pointing to API definition
	))

	return r
}

type CreateDeviceRequest struct {
	Name string `json:"name"`
}

// CreateDevice creates a new WhatsApp device session
// @Summary Create a new device
// @Description Generates a new device ID and initializes a client
// @Tags device
// @Accept json
// @Produce json
// @Param request body CreateDeviceRequest true "Device Name"
// @Success 200 {object} map[string]string
// @Router /device [post]
func (h *Handler) CreateDevice(w http.ResponseWriter, r *http.Request) {
	var req CreateDeviceRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if req.Name == "" {
		req.Name = "Instance-" + time.Now().Format("150405")
	}

	deviceID, _, err := h.manager.NewClientWithName(req.Name)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"device_id": deviceID,
		"name":      req.Name,
	})
}

// ListDevices returns a list of all connected devices with their details
// @Summary List devices
// @Description Returns a list of all active instances
// @Tags device
// @Produce json
// @Success 200 {array} whatsapp.Instance
// @Router /device [get]
func (h *Handler) ListDevices(w http.ResponseWriter, r *http.Request) {
	instances := h.manager.ListInstances()

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(instances)
}

// RenameDevice renames an existing instance
// @Summary Rename a device
// @Description Updates the name of an existing device/instance
// @Tags device
// @Accept json
// @Produce json
// @Param id path string true "Device ID"
// @Param request body CreateDeviceRequest true "New Device Name"
// @Success 200 {object} map[string]string
// @Router /device/{id}/rename [put]
func (h *Handler) RenameDevice(w http.ResponseWriter, r *http.Request) {
	deviceID := chi.URLParam(r, "id")
	var req CreateDeviceRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if req.Name == "" {
		http.Error(w, "Name is required", http.StatusBadRequest)
		return
	}

	err := h.manager.RenameInstance(deviceID, req.Name)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "success", "device_id": deviceID, "name": req.Name})
}

// GetQR returns the QR code stream for a device
// @Summary Get QR code stream
// @Description Connects to SSE stream to receive QR codes for pairing
// @Tags device
// @Param id path string true "Device ID"
// @Produce text/event-stream
// @Router /device/{id}/qr [get]
func (h *Handler) GetQR(w http.ResponseWriter, r *http.Request) {
	deviceID := chi.URLParam(r, "id")
	logger.Info("SSE connection initiated", zap.String("device", deviceID))

	// Pass the request context so we know when client disconnects
	qrChan, err := h.manager.GetQR(r.Context(), deviceID)
	if err != nil {
		logger.Error("Failed to get QR channel", zap.Error(err))
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	// Server Sent Events (SSE) for QR updates
	w.Header().Set("Content-Type", "text/event-stream")
	w.Header().Set("Cache-Control", "no-cache")
	w.Header().Set("Connection", "keep-alive")
	// CORS headers just in case, though usually handled by middleware
	w.Header().Set("Access-Control-Allow-Origin", "*")

	flusher, ok := w.(http.Flusher)
	if !ok {
		http.Error(w, "Streaming unsupported!", http.StatusInternalServerError)
		return
	}

	// Send initial ping or comment to establish connection immediately
	fmt.Fprintf(w, ": connected\n\n")
	flusher.Flush()

	for {
		select {
		case <-r.Context().Done():
			logger.Info("SSE client disconnected", zap.String("device", deviceID))
			return
		case code, ok := <-qrChan:
			if !ok {
				logger.Info("QR channel closed", zap.String("device", deviceID))
				return
			}
			if code == "" {
				continue
			}
			logger.Info("Sending QR Code event to client", zap.String("device", deviceID))
			// Using "data: <code_content>\n\n" format
			_, err := fmt.Fprintf(w, "data: %s\n\n", code)
			if err != nil {
				logger.Error("Failed to write to ResponseWriter", zap.Error(err))
				return
			}
			flusher.Flush()
		}
	}
}

// ReconnectDevice reconnects an offline instance without a QR code
// @Summary Reconnect device
// @Description Reconnects a previously paired device
// @Tags device
// @Produce json
// @Param id path string true "Device ID"
// @Success 200 {object} map[string]string
// @Router /device/{id}/reconnect [post]
func (h *Handler) ReconnectDevice(w http.ResponseWriter, r *http.Request) {
	deviceID := chi.URLParam(r, "id")
	
	err := h.manager.Connect(r.Context(), deviceID)
	if err != nil {
		logger.Error("Failed to reconnect device", zap.Error(err), zap.String("device", deviceID))
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status": "success",
	})
}

// DeleteDevice deletes a device completely
// @Summary Delete device
// @Description Deletes an instance, terminates connection and drops metadata
// @Tags device
// @Produce json
// @Param id path string true "Device ID"
// @Success 200 {object} map[string]string
// @Router /device/{id} [delete]
func (h *Handler) DeleteDevice(w http.ResponseWriter, r *http.Request) {
	deviceID := chi.URLParam(r, "id")
	
	err := h.manager.DeleteInstance(deviceID)
	if err != nil {
		logger.Error("Failed to delete device", zap.Error(err), zap.String("device", deviceID))
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status": "success",
	})
}

// GetLogs returns the recent system logs
// @Summary Get system logs
// @Description Returns the latest logs stored in memory
// @Tags app
// @Produce json
// @Success 200 {array} logger.LogEntry
// @Router /logs [get]
func (h *Handler) GetLogs(w http.ResponseWriter, r *http.Request) {
	logs := logger.GetRecentLogs()
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(logs)
}

// GetStats returns usage statistics
// @Summary Get statistics
// @Description Returns message volume stats per hour
// @Tags app
// @Produce json
// @Param id query string false "Device ID to filter by"
// @Success 200 {array} whatsapp.MessageStatGroup
// @Router /stats [get]
func (h *Handler) GetStats(w http.ResponseWriter, r *http.Request) {
    deviceID := r.URL.Query().Get("id")
	stats, err := h.manager.GetMessageStats(deviceID)
    if err != nil {
        logger.Error("Failed to fetch stats", zap.Error(err))
        http.Error(w, err.Error(), http.StatusInternalServerError)
        return
    }

	if stats == nil {
		stats = []whatsapp.MessageStatGroup{}
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(stats)
}

// GetStatus returns the current connection status of a device
// @Summary Get device status
// @Description Returns the connection status of the device
// @Tags device
// @Param id path string true "Device ID (JID)"
// @Produce json
// @Success 200 {object} map[string]string
// @Router /device/{id}/status [get]
func (h *Handler) GetStatus(w http.ResponseWriter, r *http.Request) {
	deviceID := chi.URLParam(r, "id")
	// Try direct map access first (if using JID format from URL)
	client := h.manager.GetClient(deviceID)

	if client == nil {
		// Fallback or error
		http.Error(w, "Device not found", http.StatusNotFound)
		return
	}

	status := "disconnected"
	if client.IsConnected() {
		status = "connected"
	}

	json.NewEncoder(w).Encode(map[string]string{"status": status})
}

// SendInteractiveButton sends a WhatsApp interactive button message via Queue
// @Summary Send an interactive button message
// @Description Sends an interactive button message (Queued)
// @Tags message
// @Accept json
// @Produce json
// @Param request body SendMessageRequest true "Message Request"
// @Success 200 {object} map[string]string
// @Router /message/interactive/button [post]
func (h *Handler) SendInteractiveButton(w http.ResponseWriter, r *http.Request) {
	var req SendMessageRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	payloadBytes, err := json.Marshal(req)
	if err != nil {
		http.Error(w, "Failed to encode message to queue", http.StatusInternalServerError)
		return
	}

	err = h.rmqClient.Publish(r.Context(), "send_message_queue", payloadBytes)
	if err != nil {
		http.Error(w, "Failed to enqueue message", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "queued"})
}

// SendText sends a plain text WhatsApp message via Queue
// @Summary Send a text message
// @Description Sends a plain text message (Queued)
// @Tags message
// @Accept json
// @Produce json
// @Param request body SendMessageRequest true "Message Request"
// @Success 200 {object} map[string]string
// @Router /message/send [post]
func (h *Handler) SendText(w http.ResponseWriter, r *http.Request) {
	var req SendMessageRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// Forcing type to empty so builder creates a text message
	req.Type = ""

	payloadBytes, err := json.Marshal(req)
	if err != nil {
		http.Error(w, "Failed to encode message to queue", http.StatusInternalServerError)
		return
	}

	err = h.rmqClient.Publish(r.Context(), "send_message_queue", payloadBytes)
	if err != nil {
		http.Error(w, "Failed to enqueue message", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "queued"})
}

// SendMedia sends a media WhatsApp message via Queue
// @Summary Send a media message (image, video, document, audio)
// @Description Sends a media message (Queued)
// @Tags message
// @Accept json
// @Produce json
// @Param request body SendMediaRequest true "Media Message Request"
// @Success 200 {object} map[string]string
// @Router /message/media [post]
func (h *Handler) SendMedia(w http.ResponseWriter, r *http.Request) {
	var req SendMediaRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	payloadBytes, err := json.Marshal(req)
	if err != nil {
		http.Error(w, "Failed to encode message to queue", http.StatusInternalServerError)
		return
	}

	err = h.rmqClient.Publish(r.Context(), "send_media_queue", payloadBytes)
	if err != nil {
		http.Error(w, "Failed to enqueue message", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "queued"})
}

// SendInteractiveCopyButton sends a WhatsApp interactive copy button message via Queue
// @Summary Send an interactive copy button (eg. Copy PIX)
// @Description Sends an interactive button message specifically designed to copy text (Queued)
// @Tags message
// @Accept json
// @Produce json
// @Param request body SendCopyButtonRequest true "Copy Button Request"
// @Success 200 {object} map[string]string
// @Router /message/interactive/copy [post]
func (h *Handler) SendInteractiveCopyButton(w http.ResponseWriter, r *http.Request) {
	var req SendCopyButtonRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	payloadBytes, err := json.Marshal(req)
	if err != nil {
		http.Error(w, "Failed to encode message to queue", http.StatusInternalServerError)
		return
	}

	// Reusing the standard send_message_queue as it's lightweight JSON
	err = h.rmqClient.Publish(r.Context(), "send_message_queue", payloadBytes)
	if err != nil {
		http.Error(w, "Failed to enqueue message", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{"status": "queued"})
}